Change headingMarkSpacing variable from int to float to fix a minor alignment error in the azimuth degree labels on top of the mousebox. Thanks to Packman for noticing the problem and providing the fix.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@22020 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/HUDRenderer.cxx
@@ -159,7 +159,7 @@ void			HUDRenderer::resize(bool firstTime)
   }
 
   // initialize readout spacings
-  headingMarkSpacing = (int)(5.0f * float(maxMotionSize) / headingOffset);
+  headingMarkSpacing = 5.0f * float(maxMotionSize) / headingOffset;
   altitudeMarkSpacing = floorf(5.0f * float(maxMotionSize) / altitudeOffset);
 
   setBigFontSize(w, vh);
